+++ common/types/OffsetBaseProps.ts
@@ -1,3 +1,0 @@
-import type { UseSpacingStylesProps } from './UseSpacingStylesProps';
-
-export type OffsetBaseProps = Omit<UseSpacingStylesProps, 'isInverted'>;