+++ common/types/SpacingProps.ts
@@ -1,5 +1,15 @@
 import type { SpacingScale } from './SpacingScale';
 
+export interface OffsetProps {
+  offset?: SpacingScale;
+  offsetBottom?: SpacingScale;
+  offsetEnd?: SpacingScale;
+  offsetHorizontal?: SpacingScale;
+  offsetStart?: SpacingScale;
+  offsetTop?: SpacingScale;
+  offsetVertical?: SpacingScale;
+}
+
 export interface SpacingProps {
   spacing?: SpacingScale;
   spacingBottom?: SpacingScale;