+++ mobile/hooks/useOffsetStyles.ts
@@ -1,0 +1,25 @@
+import { OffsetProps } from '@cbhq/cds-common';
+import { ViewStyle } from 'react-native';
+
+import { useInternalSpacingStyles } from './internal/useInternalSpacingStyles';
+
+export const useOffsetStyles = ({
+  offset,
+  offsetBottom,
+  offsetEnd,
+  offsetHorizontal,
+  offsetStart,
+  offsetTop,
+  offsetVertical,
+}: OffsetProps): ViewStyle => {
+  return useInternalSpacingStyles({
+    all: offset,
+    bottom: offsetBottom,
+    end: offsetEnd,
+    horizontal: offsetHorizontal,
+    start: offsetStart,
+    top: offsetTop,
+    vertical: offsetVertical,
+    isInverted: true,
+  });
+};